It works!
After a lot of painful trials and errors. This now finally works.
I've set the red "Reeling Tongue" to not launch you toward its anchor point in the scripts' inspector, that is why it behaves differently than the blue "Grapple Tongue". It cannot actually reel things in as of now.